+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.4] - 2022-12-16
+
 ### Added
 
 - New dependencies: `ts-node` and `ts-node-dev`
@@ -177,7 +179,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to change some localization strings via commands
 - A placeholder when trying to write a private message to a bot by a non-creator
 
-[Unreleased]: https://github.com/SecondThundeR/anti-premium-stickers-bot/compare/v2.0.3...HEAD
+[Unreleased]: https://github.com/SecondThundeR/anti-premium-stickers-bot/compare/v2.0.4...HEAD
+[2.0.4]: https://github.com/SecondThundeR/anti-premium-stickers-bot/compare/v2.0.3...v2.0.4
 [2.0.3]: https://github.com/SecondThundeR/anti-premium-stickers-bot/compare/v2.0.2...v2.0.3
 [2.0.2]: https://github.com/SecondThundeR/anti-premium-stickers-bot/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/SecondThundeR/anti-premium-stickers-bot/compare/v2.0.0...v2.0.1